requirement.txt
autogen-agentchat==0.6.4
autogen-core==0.6.4
autogen-ext==0.6.4
fastapi==0.116.1
starlette==0.47.2
pydantic==2.11.7
uvicorn==0.35.0
openai==1.97.0
tiktoken==0.9.0
mcp==1.12.0
fastagency[autogen,mesop,hats,fastapi,server]